+++ web/src/components/base/drawer/index.tsx
@@ -1,16 +1,5 @@
-import { useOnClickOutside } from "@hemilabs/react-hooks/useOnClickOutside";
-import { useOnKeyUp } from "@hemilabs/react-hooks/useOnKeyUp";
-import {
-  useAccountModal,
-  useChainModal,
-  useConnectModal,
-} from "@rainbow-me/rainbowkit";
-import {
-  type ReactNode,
-  type TransitionEvent,
-  useEffect,
-  useState,
-} from "react";
+import { useOverlay } from "hooks/useOverlay";
+import { type ReactNode, useEffect } from "react";
 import { createPortal } from "react-dom";
 
 type Props = {
@@ -20,57 +9,17 @@ type Props = {
 };
 
 export function Drawer({ children, onClose, requestClose }: Props) {
-  const [isOpen, setIsOpen] = useState(false);
-  const { accountModalOpen } = useAccountModal();
-  const { chainModalOpen } = useChainModal();
-  const { connectModalOpen } = useConnectModal();
-
-  function handleClose() {
-    // Prevent closing when a RainbowKit modal is open
-    if (accountModalOpen || chainModalOpen || connectModalOpen) {
-      return;
-    }
-    setIsOpen(false);
-  }
-
-  const drawerRef = useOnClickOutside<HTMLDivElement>(handleClose);
-
-  useOnKeyUp(function (e) {
-    if (e.key === "Escape") {
-      handleClose();
-    }
-  });
-
-  useEffect(
-    // Triggers the opening animation by delaying state change to the next frame.
-    // This ensures the drawer starts in the closed position (translate-x-full)
-    // and then transitions to open (translate-x-0), enabling the CSS transition.
-    function openOnMount() {
-      const rafId = requestAnimationFrame(() => setIsOpen(true));
-      return () => cancelAnimationFrame(rafId);
-    },
-    [],
-  );
+  const { handleClose, handleTransitionEnd, isOpen, ref } =
+    useOverlay<HTMLDivElement>(onClose);
 
   useEffect(
     function closeFromParent() {
       if (requestClose) {
-        setIsOpen(false);
+        handleClose();
       }
     },
     [requestClose],
   );
-
-  function handleTransitionEnd(e: TransitionEvent) {
-    // Ignore bubbled events from child transitions
-    if (e.target !== e.currentTarget) {
-      return;
-    }
-    if (!isOpen) {
-      // Unmount after slide-out animation completes
-      onClose();
-    }
-  }
 
   return createPortal(
     <>
@@ -80,7 +29,7 @@ export function Drawer({ children, onClose, requestClose }: Props) {
       <div
         className={`fixed inset-y-0 right-0 z-20 w-full rounded-l-lg bg-white shadow-xl transition-transform duration-300 md:w-md ${isOpen ? "translate-x-0" : "translate-x-full"}`}
         onTransitionEnd={handleTransitionEnd}
-        ref={drawerRef}
+        ref={ref}
       >
         {children}
       </div>
